refactor typeutils
